requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
python-docx==1.1.0
striprtf==0.0.26
lxml==5.2.2
reportlab==4.2.0
sse-starlette==2.1.0
urllib3==2.2.2
